+++ src/test/java/com/epam/reportportal/utils/ParameterUtilsTest.java
@@ -18,6 +18,7 @@ package com.epam.reportportal.utils;
 
 import com.epam.reportportal.annotations.ParameterKey;
 import com.epam.reportportal.service.item.TestCaseIdEntry;
+import com.epam.reportportal.utils.markdown.MarkdownUtilsTest;
 import com.epam.ta.reportportal.ws.model.ParameterResource;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.Test;
@@ -234,5 +235,22 @@ public class ParameterUtilsTest {
 			assertThat(p.getKey(), equalTo(PARAM_KEY_VALUES.get(i).getKey()));
 			assertThat(p.getValue(), equalTo(v == null ? "NULL" : v.toString()));
 		});
+	}
+
+	@Test
+	public void test_parameters_format() {
+		ParameterResource varA = new ParameterResource();
+		ParameterResource varB = new ParameterResource();
+		ParameterResource result = new ParameterResource();
+
+		varA.setKey("var_a");
+		varB.setKey("var_b");
+		result.setKey("result");
+		varA.setValue("2");
+		varB.setValue("2");
+		result.setValue("4");
+
+		assertThat(ParameterUtils.formatParametersAsTable(Arrays.asList(varA, varB, result)),
+				equalTo(MarkdownUtilsTest.ONE_ROW_EXPECTED_TABLE));
 	}
 }